requirements.txt
catalogue==2.0.8
gensim==3.8.3
image==1.5.33
matplotlib==3.5.1
multiprocess==0.70.12.2
networkx==2.6.3
nltk==3.7
numpy==1.21.6
pandas==1.3.5
pdf2image==1.16.0
Pillow==9.5.0
pytesseract==0.3.10
requests==2.27.1
spacy==2.3.2
tomotopy==0.12.4
tqdm==4.64.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
